+++ src/components/forms/ComboBox/ComboBox.stories.tsx
@@ -71,8 +71,14 @@ export const withLabel = (): React.ReactElement => {
 
   return (
     <Form onSubmit={noop}>
-      <Label htmlFor="fruit">Select A Fruit</Label>
-      <ComboBox id="fruit" name="fruit" options={fruitList} onChange={noop} />
+      <Label htmlFor="fruit">Select a fruit</Label>
+      <ComboBox
+        id="fruit"
+        name="fruit"
+        options={fruitList}
+        onChange={noop}
+        ulProps={{ 'aria-labelledby': 'fruit-label' }}
+      />
     </Form>
   )
 }