+++ src/index.ts
@@ -13,6 +13,17 @@ export { Tag } from './components/Tag/Tag'
 export { Tooltip } from './components/Tooltip/Tooltip'
 export { SideNav } from './components/SideNav/SideNav'
 
+/** Collection components */
+export { Collection } from './components/Collection/Collection'
+export { CollectionItem } from './components/Collection/CollectionItem'
+export { CollectionHeading } from './components/Collection/CollectionHeading'
+export { CollectionDescription } from './components/Collection/CollectionDescription'
+export { CollectionMeta } from './components/Collection/CollectionMeta'
+export { CollectionMetaItem } from './components/Collection/CollectionMetaItem'
+export { CollectionMetaItemTag } from './components/Collection/CollectionMetaItemTag'
+export { CollectionThumbnail } from './components/Collection/CollectionThumbnail'
+export { CollectionCalendarDate } from './components/Collection/CollectionCalendarDate'
+
 /** Grid components */
 export { GridContainer } from './components/grid/GridContainer/GridContainer'
 export { Grid } from './components/grid/Grid/Grid'
@@ -36,6 +47,7 @@ export { Radio } from './components/forms/Radio/Radio'
 export { RangeInput } from './components/forms/RangeInput/RangeInput'
 export { Textarea } from './components/forms/Textarea/Textarea'
 export { TextInput } from './components/forms/TextInput/TextInput'
+export { TimePicker } from './components/forms/TimePicker/TimePicker'
 export { ValidationChecklist } from './components/forms/Validation/ValidationChecklist'
 export { ValidationItem } from './components/forms/Validation/ValidationItem'
 
@@ -54,10 +66,12 @@ export { Title } from './components/header/Title/Title'
 /** Identifier Components */
 export { Identifier } from './components/Identifier/Identifier/Identifier'
 export { IdentifierGov } from './components/Identifier/IdentifierGov/IdentifierGov'
+export { IdentifierIdentity } from './components/Identifier/IdentifierIdentity/IdentifierIdentity'
 export { IdentifierLink } from './components/Identifier/IdentifierLink/IdentifierLink'
 export { IdentifierLinkItem } from './components/Identifier/IdentifierLinkItem/IdentifierLinkItem'
 export { IdentifierLinks } from './components/Identifier/IdentifierLinks/IdentifierLinks'
 export { IdentifierLogo } from './components/Identifier/IdentifierLogo/IdentifierLogo'
+export { IdentifierLogos } from './components/Identifier/IdentifierLogos/IdentifierLogos'
 export { IdentifierMasthead } from './components/Identifier/IdentifierMasthead/IdentifierMasthead'
 
 /** Footer components */
@@ -68,7 +82,7 @@ export { FooterNav } from './components/Footer/FooterNav/FooterNav'
 export { Logo } from './components/Footer/Logo/Logo'
 export { SocialLinks } from './components/Footer/SocialLinks/SocialLinks'
 
-/** Card Components */
+/** Card components */
 export { CardGroup } from './components/card/CardGroup/CardGroup'
 export { Card } from './components/card/Card/Card'
 export { CardHeader } from './components/card/CardHeader/CardHeader'
@@ -81,7 +95,20 @@ export { BreadcrumbBar } from './components/breadcrumb/BreadcrumbBar/BreadcrumbB
 export { Breadcrumb } from './components/breadcrumb/Breadcrumb/Breadcrumb'
 export { BreadcrumbLink } from './components/breadcrumb/BreadcrumbLink/BreadcrumbLink'
 
+/** StepIndicator components */
+export { StepIndicator } from './components/stepindicator/StepIndicator/StepIndicator'
+export { StepIndicatorStep } from './components/stepindicator/StepIndicatorStep/StepIndicatorStep'
+
 export { Search } from './components/Search/Search'
+
+export { SummaryBox } from './components/SummaryBox/SummaryBox'
+
+/** ProcessList components */
+export { ProcessList } from './components/ProcessList/ProcessList/ProcessList'
+export { ProcessListItem } from './components/ProcessList/ProcessListItem/ProcessListItem'
+export { ProcessListHeading } from './components/ProcessList/ProcessListHeading/ProcessListHeading'
+
+export { SiteAlert } from './components/SiteAlert/SiteAlert'
 
 /** Truss-designed components */
 export {